[BE] feat: Bean Validation 적용 (#15)
* feat: 피드백 추가 API validation 추가(#12)

request, response 일부 수정 및 Bean Validation 추가

* feat: 피드백 수정 API validation 추가(#12)

* feat: 댓글 추가, 수정 API validation 추가(#12)

* feat: 친구 요청, 수락 API validation 추가(#12)

* feat: 예상 질문 추가, 수정 API validation 추가(#12)

* feat: 이력서 수정 API validation 추가(#12)

* chore: response 타입 변경(#12)

* chore: 일관성 있도록 변수명 수정(#12)

* feat: 타입 수정 및 response 데이터 수정(#12)

일관성있도록 타입 수정 및 회의 내용에 따라 response 데이터 수정

* feat: 일관성 있도록 타입 변경 및 예상 질문 response 일부 수정(#12)

회의 내용에 따라 예상 질문 목록 응답 시 labelId 대신 labenContent를 응답하도록 함

* feat: 타입 수정 및 response 수정(#12)

일관성 있도록 타입 수정 및 회의 내용에 따라 피드백 수정 API response 수정

* chore: 댓글 API 타입 수정(#12)

* chore: 전체 객체 클래스 타입 수정(#12)

* feat: 수정된 API 동작 테스트(#12)

* feat: 친구 삭제(또는 친구 요청 거절) API 문서 작성(#12)"

회의 내용에 따라 추가 API 문서 작성

+++ src/main/java/reviewme/be/feedback/entity/Feedback.java
@@ -17,7 +17,7 @@ public class Feedback {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Long id;
 
     @OneToOne
     @JoinColumn(name = "user_id")
@@ -36,9 +36,9 @@ public class Feedback {
     private Resume resume;
 
     private String content;
-    private int resumePage;
+    private Integer resumePage;
     private Boolean checked;
-    private int childCnt;
+    private Long childCnt;
     private LocalDateTime createdAt;
     private LocalDateTime deletedAt;
 }
